requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
clamd==1.0.2
Click==7.0
configobj==5.0.6
configparser==4.0.2
decorator==4.4.1
Django==3.0.2
django-clamd==0.4.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
etelemetry==0.1.2
filelock==3.0.12
fitz==0.0.1.dev2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
isodate==0.6.0
jdcal==1.4.1
lxml==4.5.0
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.1
nipype==1.4.2
numpy==1.18.1
openpyxl==3.0.3
packaging==20.1
pandas==1.0.1
Pillow==7.1.1
prov==1.5.3
pydot==1.4.1
pydotplus==2.0.2
PyJWT==1.7.1
PyMuPDF==1.16.10
pyparsing==2.4.6
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-http-client==3.2.5
pytz==2019.3
pyxnat==1.2.1.0.post3
razorpay==1.2.0
rdflib==4.2.2
reportlab==3.5.34
requests==2.22.0
scipy==1.4.1
sendgrid==6.1.2
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
traits==6.0.0
urllib3==1.25.8
